+++ Source/Sitecore.FakeDb/Data/Items/ItemHelper.cs
@@ -2,7 +2,6 @@
 {
   using Sitecore.Data;
   using Sitecore.Data.Items;
-  using Sitecore.Data.Managers;
   using Sitecore.Diagnostics;
   using Sitecore.Globalization;